requirements.txt
asgiref==3.7.2
astroid==3.0.1
autopep8==2.0.4
colorama==0.4.6
dill==0.3.7
Django==4.2.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
isort==5.12.0
mccabe==0.7.0
platformdirs==3.11.0
pycodestyle==2.11.1
pylint==3.0.2
pytz==2024.1
sqlparse==0.4.4
tomlkit==0.12.1
tzdata==2023.3
